requirements.txt
antiorm==1.2.1
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
idna==2.10
numpy==1.19.0
pandas==1.0.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
tornado==6.0.4
urllib3==1.25.9
